requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
email-validator==2.1.0.post1
imaplib2==3.6
beautifulsoup4==4.10.0
numpy>=1.26.0
pandas>=2.0.0 